Fix compile errors with MacOS - _POSIX_C_SOURCE must not be defined for num_CPU()
