+++ client/cmdhfepa.c
@@ -16,6 +16,7 @@
 #include <unistd.h>
 #include <stdio.h>
 #include "util.h"
+#include "util_posix.h"
 #include "proxmark3.h"
 #include "ui.h"
 #include "cmdparser.h"
