+++ client/cmdhfmfhard.c
@@ -28,6 +28,7 @@
 #include "cmdmain.h"
 #include "ui.h"
 #include "util.h"
+#include "util_posix.h"
 #include "crapto1/crapto1.h"
 #include "parity.h"
 #include "hardnested/hardnested_bruteforce.h"
